+++ src/WeihanLi.Common.Logging.Log4Net/Log4NetLogHelperLogger.cs
@@ -43,7 +43,7 @@ namespace WeihanLi.Common.Logging.Log4Net
             }
         }
 
-        public void Log(LogHelperLevel loggerLevel, string message, Exception exception)
+        public void Log(LogHelperLevel loggerLevel, Exception exception, string message)
         {
             if (IsEnabled(loggerLevel))
             {