+++ src/WeihanLi.Common/Logging/MicrosoftLoggingLogHelperProvider.cs
@@ -60,7 +60,7 @@ namespace WeihanLi.Common.Logging
             }
         }
 
-        public void Log(LogHelperLevel loggerLevel, string message, Exception exception)
+        public void Log(LogHelperLevel loggerLevel, Exception exception, string message)
         {
             if (IsEnabled(loggerLevel))
             {